Fix more upstream merge conflicts

+++ src/araxiaonline/mcp/ServerTools.cpp
@@ -874,7 +874,7 @@ void RegisterServerTools()
                     if (!itemTemplate) continue;
                     
                     // Quality filter - check bucket's quality mask
-                    if (qualityFilter >= 0 && itemTemplate->GetQuality() < qualityFilter)
+                    if (qualityFilter >= 0 && itemTemplate->GetQuality() < static_cast<uint32>(qualityFilter))
                         continue;
                     
                     // Name filter using bucket's FullName (same as server's BuildListBuckets)
